fix playrate, deprecate old animation player
